% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsner.R
\name{tsne.data.frame}
\alias{tsne.data.frame}
\title{tsne.data.frame implements t-Distributed Stochastic Neighbor Embedding (t-SNE) 
for data frame inputs}
\usage{
\method{tsne}{data.frame}(X, ...)
}
\arguments{
\item{X}{data.frame; Data Frame with only columns with numeric data}

\item{dims}{integer; Output dimensionality (default: 2)}

\item{initial_dims}{integer; the number of dimensions that should be retained in the initial PCA step (default: -1 (all))}

\item{perplexity}{numeric; Perplexity parameter}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy, default: 0.5)}

\item{pca}{logical; Whether an initial PCA step should be performed (default: FALSE)}

\item{max_iter}{integer; Number of iterations (default: 1000)}

\item{verbose}{logical; Whether progress updates should be printed (default: FALSE)}
}
\description{
tsne.data.frame implements t-Distributed Stochastic Neighbor Embedding (t-SNE) 
for data frame inputs
}
\examples{
\dontrun{
tsnedf <- tsne(X, dims, initial_dims, perplexity, max_iter, pca, theta, verbose)
}
}

