% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsner.R
\name{tsne.matrix}
\alias{tsne.matrix}
\title{tsne.matrix implements t-Distributed Stochastic Neighbor Embedding (t-SNE) 
for matrix inputs}
\usage{
\method{tsne}{matrix}(
  X,
  dims = 2,
  initial_dims = -1,
  perplexity = 30,
  max_iter = 100,
  pca = FALSE,
  theta = 0.6,
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix; Data matrix}

\item{dims}{integer; Output dimensionality (default: 2)}

\item{initial_dims}{integer; the number of dimensions that should be retained in the initial PCA step (default: -1 (all))}

\item{perplexity}{numeric; Perplexity parameter}

\item{max_iter}{integer; Number of iterations (default: 1000)}

\item{pca}{logical; Whether an initial PCA step should be performed (default: FALSE)}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy, default: 0.5)}

\item{verbose}{logical; Whether progress updates should be printed (default: FALSE)}
}
\description{
tsne.matrix implements t-Distributed Stochastic Neighbor Embedding (t-SNE) 
for matrix inputs
}
\examples{
\dontrun{
tsnedf <- tsne(X, dims, initial_dims, perplexity, max_iter, pca, theta, verbose)
}
}
